modify genres
