+++ src/App.js
@@ -10,6 +10,7 @@ import GenresPage from "./Components/GenresPage/GenresPage";
 
 function App() {
   const [posts, setPosts] = useState([]);
+  const [moviesGenres, setMoviesGenres] = useState([]);
 
   useEffect(() => {
     axios
@@ -19,19 +20,35 @@ function App() {
       .then((res) => {
         console.log(res);
         setPosts(res.data.results);
-        console.log("post", posts);
       });
   }, []);
+  console.log("post", posts);
+
+  const movieGenres = () => {
+    axios
+      .get(
+        "https://api.themoviedb.org/3/genre/movie/list?api_key=6a3a9e9a61085d657b30d36d1c7b5ba7"
+      )
+      .then((res) => {
+        console.log("genres", res.data.genres);
+        setMoviesGenres(res.data.genres);
+      });
+  };
+
+  useEffect(() => {
+    movieGenres();
+  }, []);
+
   return (
     <div className="App">
-      <Example />
+      <Example moviesGenres={moviesGenres} />
       <Routes>
         <Route path="/" element={<Home posts={posts} />} />
         <Route path="/movies/:type" element={<MoviePage />} />
         <Route path="/movie/:movieId" element={<Description />} />
         <Route
           path="/moviesGenres/:genres"
-          element={<GenresPage posts={posts} />}
+          element={<GenresPage moviesGenres={moviesGenres} />}
         />
       </Routes>
     </div>
